refactor(skill-selector): 将文件移至 guid 子目录，限定作用域
