+++ src/renderer/pages/guid/components/SkillSelectorSkeleton.tsx
