+++ src/renderer/pages/guid/index.tsx
@@ -10,8 +10,8 @@ import { skillHub } from '@/common/ipcBridge';
 import { useAuth } from '@/renderer/context/AuthContext';
 import { ipcBridge } from '@/common';
 import { resolveAssistantName } from '@/renderer/shared/agents/assistantAdapter';
-import SkillSelectorPopover from '@/renderer/components/SkillSelectorPopover';
-import { useSkillSelectorController, type SkillSelectorItem, stripAtQuery } from '@/renderer/hooks/useSkillSelectorController';
+import SkillSelectorPopover from '@/renderer/pages/guid/components/SkillSelectorPopover';
+import { useSkillSelectorController, type SkillSelectorItem, stripAtQuery } from '@/renderer/pages/guid/hooks/useSkillSelectorController';
 import { getInstalledSkillDisplay, resolveSkillIcon } from '@/renderer/utils/skillDisplay';
 import { useConversationTabs } from '@/renderer/pages/conversation/context/ConversationTabsContext';
 import { openExternalUrl, isElectronDesktop, resolveExtensionAssetUrl } from '@/renderer/utils/platform';
